fix: Integer.parseInt 잘못쓰던부분 Long.parseLong으로 변경

+++ src/main/java/lotto/view/InputView.java
@@ -15,12 +15,12 @@ public class InputView {
 
   public static long inputMoney() {
     System.out.println("구입금액을 입력해 주세요.");
-    return Integer.parseInt(scanner.nextLine());
+    return Long.parseLong(scanner.nextLine());
   }
 
   public static long inputManualLottoCnt(Money money) {
     System.out.println("\n수동으로 구매할 로또 수를 입력해 주세요.");
-    long manualLottoCnt = Integer.parseInt(scanner.nextLine());
+    long manualLottoCnt = Long.parseLong(scanner.nextLine());
 
     if (manualLottoCnt > LottoGame.purchaseCnt(money)) {
       throw new ExceedPurchaseCountException();
